+++ src/__stories__/config.js
@@ -1,8 +1,7 @@
 import Events from "../Utils/Events"
 import { addDecorator, configure } from "@storybook/react"
 import { withOptions } from "@storybook/addon-options"
-import { injectGlobal } from "styled-components"
-import { createMediaStyle } from "../Utils/Responsive"
+import { createMediaStyle } from "Utils/Responsive"
 
 const req = require.context("../", true, /\.story\.tsx$/)
 
@@ -10,9 +9,10 @@ function loadStories() {
   req.keys().forEach(filename => req(filename))
 }
 
-injectGlobal`
-  ${createMediaStyle()}
-`
+// Add RRM styles to Storybooks head
+const rrmStyle = document.createElement("style")
+rrmStyle.innerHTML = createMediaStyle()
+document.head.appendChild(rrmStyle)
 
 addDecorator(
   withOptions({